Several warnings fixed. enumeration convert method for RTCIceGatherer object exposed

+++ projects/api/ortc-winrt-api/ortc-winrt-api.Shared/RTCSctpTransport.h
@@ -35,7 +35,7 @@ namespace ortc_winrt_api
   public ref class RTCSctpCapabilities sealed
   {
   public:
-    property uint16 MaxMessageSize;
+    property size_t MaxMessageSize;
   };
 
   //------------------------------------------
